utils - make utils core-only, relocate gui-only utils to gui
